src/dashboard/requirements.txt
flask
flask-socketio
confluent_kafka
python-dotenv
pyyaml
eventlet
